wip: extend the types, and some clean up
